totemsrp: fix stack buffer overflow + alloca→malloc + sanity checks
Session-3 followup: 300-node simulation exposed critical bugs.

1. STACK BUFFER OVERFLOW in message_handler_orf_token (regrssion from
   session-1 fix #8): struct rtr_item is 16 bytes (ring_id=12B + seq=4B),
   not 8 as the comment claimed.  RETRANSMIT_ENTRIES_MAX=256 tokens are
   4153 bytes but token_storage/token_convert were only 4096 bytes —
   writing 253-256 RTR entries overflowed the stack by up to 57 bytes.
   Fix: raise RETRANSMIT_ENTRIES_MAX to 384 (=PROCESSOR_COUNT_MAX),
   enlarge token_storage/token_convert to TOKEN_STORAGE_BYTES=8192.
   At 300 nodes this eliminates RTR starvation entirely (sim: 0/399
   rotations starved vs 13/400 with the old value).

2. alloca→malloc in message_handler_memb_join and
   message_handler_memb_commit_token: both allocations are called with
   unvalidated msg_len from the network.  A crafted large msg_len could
   exhaust the stack.  Now uses malloc with NULL check; commit token
   allocation moved to after check_memb_commit_token_sanity so msg_len
   is bounds-checked before the allocation.

3. Integer overflow in sanity checkers: check_memb_join_sanity and
   check_memb_commit_token_sanity compute required_len from network-
   supplied entry counts.  On 32-bit platforms, a large count × struct
   size could wrap to zero, bypassing the length check.  Added
   PROCESSOR_COUNT_MAX bounds checks on proc_list_entries /
   failed_list_entries / addr_entries before the multiplication.

4. totemconfig.c format string: three snprintf calls used %u with plain
   int i (loop variable) — added (unsigned int)i casts; also fixed a
   TOTEM_PRIVATE_KEY_LEN_MIN * 8 expression passed to %u.

5. sim300.py: added --rtr-max and --window-size / --fixed CLI flags to
   allow simulation with patched protocol constants.

+++ exec/totemconfig.c
@@ -2125,20 +2125,20 @@ int totem_config_validate (
 		if (totem_config->interfaces[i].mcast_addr.family == AF_INET6 &&
 			totem_config->node_id == 0) {
 		        snprintf (local_error_reason, sizeof(local_error_reason),
-					"An IPV6 network requires that a node ID be specified for interface %u", i);
+					"An IPV6 network requires that a node ID be specified for interface %u", (unsigned int)i);
 			goto parse_error;
 		}
 
 		if (totem_config->broadcast_use == 0 && totem_config->transport_number == TOTEM_TRANSPORT_UDP) {
 			if (totem_config->interfaces[i].mcast_addr.family != totem_config->interfaces[i].bindnet.family) {
 				snprintf (local_error_reason, sizeof(local_error_reason),
-						"Multicast address family does not match bind address family for interface %u", i);
+						"Multicast address family does not match bind address family for interface %u", (unsigned int)i);
 				goto parse_error;
 			}
 
 			if (totemip_is_mcast (&totem_config->interfaces[i].mcast_addr) != 0) {
 				snprintf (local_error_reason, sizeof(local_error_reason),
-						"mcastaddr is not a correct multicast address for interface %u", i);
+						"mcastaddr is not a correct multicast address for interface %u", (unsigned int)i);
 				goto parse_error;
 			}
 		}
@@ -2247,7 +2247,7 @@ static int read_keyfile (
 	if (res < TOTEM_PRIVATE_KEY_LEN_MIN) {
 		snprintf (error_string_response, sizeof(error_string_response),
 			"Could only read %d bits of minimum %u bits from %s.\n",
-			 res * 8, TOTEM_PRIVATE_KEY_LEN_MIN * 8, key_location);
+			 res * 8, (unsigned int)(TOTEM_PRIVATE_KEY_LEN_MIN * 8), key_location);
 		goto parse_error;
 	}
